+++ data_source_firewall_rule.go
@@ -80,23 +80,27 @@ func dataSourceFirewallRuleRead(d *schema.ResourceData, m interface{}) error {
 	networkInterfaceID := d.Get(dataSourceFirewallRuleNetworkInterfaceIDKey).(string)
 	serverID := d.Get(dataSourceFirewallRuleServerIDKey).(string)
 
-	req, reqErr := clouddk.GetClientRequestObject(&clientSettings, "GET", fmt.Sprintf("cloudservers/%s/network-interfaces/%s/firewall-rules/%s", serverID, networkInterfaceID, firewallRuleID), new(bytes.Buffer))
+	req, err := clouddk.GetClientRequestObject(&clientSettings, "GET", fmt.Sprintf("cloudservers/%s/network-interfaces/%s/firewall-rules/%s", serverID, networkInterfaceID, firewallRuleID), new(bytes.Buffer))
 
-	if reqErr != nil {
-		return reqErr
+	if err != nil {
+		return err
 	}
 
 	client := &http.Client{}
-	res, resErr := client.Do(req)
+	res, err := client.Do(req)
 
-	if resErr != nil {
-		return resErr
+	if err != nil {
+		return err
 	} else if res.StatusCode != 200 {
 		return fmt.Errorf("Failed to read the information about the firewall rule - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	firewallRule := clouddk.FirewallRuleBody{}
-	json.NewDecoder(res.Body).Decode(&firewallRule)
+	err = json.NewDecoder(res.Body).Decode(&firewallRule)
+
+	if err != nil {
+		return err
+	}
 
 	return dataSourceFirewallRuleReadResponseBody(d, m, &firewallRule)
 }