+++ data_source_packages.go
@@ -38,23 +38,27 @@ func dataSourcePackages() *schema.Resource {
 // dataSourcePackagesRead reads information about server packages.
 func dataSourcePackagesRead(d *schema.ResourceData, m interface{}) error {
 	clientSettings := m.(clouddk.ClientSettings)
-	req, reqErr := clouddk.GetClientRequestObject(&clientSettings, "GET", "cloudservers/get-packages", new(bytes.Buffer))
+	req, err := clouddk.GetClientRequestObject(&clientSettings, "GET", "cloudservers/get-packages", new(bytes.Buffer))
 
-	if reqErr != nil {
-		return reqErr
+	if err != nil {
+		return err
 	}
 
 	client := &http.Client{}
-	res, resErr := client.Do(req)
+	res, err := client.Do(req)
 
-	if resErr != nil {
-		return resErr
+	if err != nil {
+		return err
 	} else if res.StatusCode != 200 {
 		return fmt.Errorf("Failed to read the information about the packages - Reason: The API responded with HTTP %s", res.Status)
 	}
 
 	list := make(clouddk.PackageeListBody, 0)
-	json.NewDecoder(res.Body).Decode(&list)
+	err = json.NewDecoder(res.Body).Decode(&list)
+
+	if err != nil {
+		return err
+	}
 
 	ids := make([]interface{}, len(list))
 	names := make([]interface{}, len(list))